Fix "shown" undefined JavaScript error
this error was caused when the JavaScript code that initializes
the images on the Home page was being executed before the actual
HTML elements were found!

This commit pages the extra_javascript block at the end of the
document, right after the closing </footer> tag and removes the
window.onload setting. Instead just interpret the JavaScript
instructions like it used to be before this code was split from
the Home page.