Add a clarification about configuration of default settings
when changing settings customer needs to know what they want so that the
Kiwi TCMS team can configure it exactly as requested. Vague requests or
unknown scenarios need to be researched & experimented with by the
customer until they are clear what the result of their request is going
to be!